Create stable_dialogs_escape_WORKINPROGRESS.xml
little work in progress of one of the XML files for translation. Most of it is AI made, but I'm working with a copy of the original Russian XML file to make the translation more accurate and see which parts should be changed, hopefully only partially, as the point isn't to change intent, but rather to reword the AI translation in contrast with the Russian original to see what would fit best.